+++ api/src/main/java/org/allaymc/api/client/skin/PersonaPieceTint.java
@@ -1,9 +1,0 @@
-package org.allaymc.api.client.skin;
-
-import java.util.List;
-
-/**
- * @author LucGamesYT | daoge_cmd
- */
-public record PersonaPieceTint(String pieceType, List<String> colors) {
-}