+++ api/src/main/java/org/allaymc/api/client/skin/SkinAnimation.java
@@ -1,7 +1,0 @@
-package org.allaymc.api.client.skin;
-
-/**
- * @author LucGamesYT | daoge_cmd
- */
-public record SkinAnimation(Image image, int type, float frames, int expression) {
-}